Add catalog scans for bucket ranges
This is an alternate way of processing packages. Before this change, the only way that packages could be processed by
drivers is by starting a catalog scan with a min and max catalog commit timestamp range. This new mode of processing
bucket ranges allows us to reprocess a subset of all discovered packages. The bucket ranges refer to partitions of all
packages known by the LoadBucketedPackage driver. The packages are partitioned into 1000 partitions and you reprocess
ranges of 1 out of a 1000 indexes.