Move several tests to use Verify snapshot library
* CatalogScanServiceIntegrationTest
* PackageCertificateToCsvIntegrationTest
* CleanupOrphanRecordsServiceTest
* CatalogRangeSameAsBucketRange

This does not use the object serialization in Verify and instead asserts the existing
test data strings as-is. Table entities and wide entities are serialized to JSON then
handed off to Verify. CSVs are left as a giant CSV string then handed off to Verify.
JSON and CSV are removed as "text" file types in Verify's text file look-up table so
that existing encoding decisions are preserved. This makes the existing data files just
move to another part of the source tree without any diff, reducing the change of regression.