Merged PR 607776: Add Title property to timers so GitHubUsageToCsv is correct
Using Humanizer library, `GitHubUsageToCsv` appears as `Git hub usage to csv`. There is an existing fix for `csv` to `CSV` but `Git hub` should be `GitHub`. This PR adds a `Title` property to timers so that we can set this title manually and use a proper `GitHub usage to CSV`.

Before:
![image.png](https://dev.azure.com/devdiv/0bdbc590-a062-4c3f-b0f6-9383f67865ee/_apis/git/repositories/83c36752-0bff-40ba-ae91-d15647d86978/pullRequests/607776/attachments/image.png)

After:
![image (2).png](https://dev.azure.com/devdiv/0bdbc590-a062-4c3f-b0f6-9383f67865ee/_apis/git/repositories/83c36752-0bff-40ba-ae91-d15647d86978/pullRequests/607776/attachments/image%20%282%29.png)

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request adds a `Title` property to timers to ensure the `GitHubUsageToCsv` functionality is correct.
- Added `Title` property to `ITimer` interface in `src/Logic/Timers/ITimer.cs`.
- Updated `TimerCard.razor` to use the new `Title` property.
- Modified `TimerState` and `SpecificTimerExecutionService` to include the `Title` property.
- Implemented `Title` property in various timer classes such as `CatalogScanUpdateTimer`, `AuxiliaryFileUpdaterTimer`, and `GitHubUsageToCsvUpdater`.