Merged PR 611316: Add bucket dimension to metrics for debugging
Both USNC and CUS instances of NuGet Insights got stuck on PackageArchiveToCsv. Some data appears to get dropped from a single CSV file. This was detected in the Kusto validation step.

In USNC region, it was 1041 package identities in bucket 47.
In CUS region, it was 1956 package identities in bucket 210.

The problem happened sometime after 2025-02-12T20:02:13.8042199Z in USNC.
The problem happened sometime after 2025-02-13T22:59:07.1697393Z in CUS.

It appears on the first attempt at PackageArchiveToCsv some data got dropped from the CSV and subsequent attempts at PackageArchiveToCsv failed to recover the missing records (because the cursor for the driver had moved forward)

In USNC (bucket 47), it appears records related to `uno.extensions.authentication.ui/2.6.0-dev.27` to `zzzlab.office/0.2123.1320.2601` went missing. This looks like the end of the CSV file.

In CUS (bucket 210), it appears records related to `strawberryshake.blazor/14.0.0-p.98` to `соstura.fоdу/5.8.0-alpha0098` (non-ASCII characters -- putting it at the end of the file) went missing. This too looks like the end of the CSV file.

The root cause is unknown. Records went missing despite them being unrelated to new catalog entries. For CUS, The latest commit timestamp seen for a missing package identity was `2025-02-13T09:32:04.7881075Z`.  This commit timestamp is more than 6 hours before the problem, meaning that this is not related to newly produced records and is instead related to handling existing records in the CSV. This points to a problem in `CsvRecordStorageService`, particularly the "big mode" flow. There is perhaps an edge case where the end of the file is dropped during pruning/merging or perhaps the existing file is not fully read. This telemetry should help us figure it out.

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
Added a bucket dimension to metrics for enhanced debugging capabilities.
- `src/Worker.Logic/CsvRecordStorage/CsvRecordStorageService.cs`: Added "Bucket" dimension to various telemetry metrics.
- `src/Worker.Logic/AppendResults/AppendResultStorageService.cs`: Included "Bucket" dimension in append metrics.
- `test/Worker.Logic.Test/Drivers/PackageAssetToCsv/PackageAssetToCsvIntegrationTest.cs`: Updated tests to reflect the new "Bucket" dimension in metrics.
- `test/Worker.Logic.Test/AppendResults/AppendResultStorageServiceTest.cs`: Adjusted tests to include the "Bucket" dimension in metric assertions.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->