Merged PR 623647: Plumb TimeProvider into memory storage, isolate memory store
This is to allow [TimeProvider](https://learn.microsoft.com/en-us/dotnet/standard/datetime/timeprovider-overview) mocking to be used in tests, allowing for more precise testing. This also isolates the in-memory Azure Storage implementation to a single test. Currently the tests are isolated by container name but this should not be needed for in-memory storage.

We have several tests that sleep, waiting for a lease to expire. This can be fixed with TimeProvider.

----
#### AI description  (iteration 1)
#### PR Classification
Code cleanup and refactoring

#### PR Summary
This pull request integrates `TimeProvider` into memory storage and refactors the memory store to improve testability and maintainability.
- `ServiceClientFactory` updated to include `TimeProvider` and memory stores.
- `MemoryBlobServiceClient`, `MemoryQueueServiceClient`, and `MemoryTableServiceClient` constructors now require `TimeProvider`.
- Refactored `BaseLogicIntegrationTest` to initialize `TimeProvider` and memory stores.
- Updated `MemoryBlobServiceStore`, `MemoryQueueServiceStore`, and `MemoryTableServiceStore` to use `TimeProvider`.
- Refactored test classes in `StorageLeaseServiceTest` to inherit from `StorageLeaseServiceTest` and use `TimeProvider`.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->