Merged PR 629982: Add FileSystemHttpCacheHandler to cache HTTP responses
This would allow us to run tests offline. And not connect to nuget.org which violates CFS guidelines.

It is currently disabled. A future PR will populate the cache and enable it in the CI and locally under TBD conditions.

----
New feature: Introduces a handler to cache HTTP responses on the file system.

This pull request adds a new `FileSystemHttpCacheHandler` class to cache HTTP responses, improving efficiency by reducing redundant network requests.
- `FileSystemHttpCacheHandler.cs`: Implements caching logic with various modes (ReadAndWriteOnlyMissing, WriteOnlyMissing, WriteAlways, ReadOnly) and handles sensitive headers, request/response caching, and file operations for cache management.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->