Merged PR 630054: Cache container initialization
Currently many services have an `InitializeAsync()` method which calls dependency `InitializeAsync()`. At the end of the call chain is storage initialization, such as creating blob containers, queues, or tables. Creating these can be cached for process lifetime.

The only complication is the `DestroyAsync()` methods which previously just deleting the corresponding containers. To ensure existing in-memory caches are still viable, `DestroyAsync()` now deletes _and then recreates_ the container.

There is a time limit on the cache to avoid permenant mismatch between `created = true` being cached and the container not existing for whatever reason.

----
New feature

This pull request introduces a caching mechanism for container initialization to improve performance and reliability.
- Added `ContainerInitializationState` class in `src/Logic/AzureStorage/ContainerInitializationState.cs` to handle container initialization and destruction.
- Refactored `InitializerHostedService` in `src/Website/HostedServices/InitializerHostedService.cs` to use the new container initialization state.
- Updated `WorkerQueueFactory` in `src/Worker.Logic/Queues/WorkerQueueFactory.cs` to utilize the new container initialization state for queue management.
- Modified `AuxiliaryFileUpdaterService` in `src/Worker.Logic/MessageProcessors/AuxiliaryFileUpdater/AuxiliaryFileUpdaterService.cs` to integrate the new initialization state.
- Added new end-to-end tests in `test/Worker.Test/EndToEnd/CanRunTimers.cs` and deleted obsolete tests.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->