Merged PR 632538: Use package filter when pruning CSV records
This is to mitigate an issue where unicode packages are not processed properly on NuGet.org. We are ignoring them right now. This PR uses the package filter in one more place: the service the creates the CSV files (`CsvRecordStorageService`). This allows CSV records to be pruned if the package filter applies to them, even if they were written to the CSV file previously (before the filter was configured).

----
Enhancement: Introduces a package filter to prune CSV records effectively.

This pull request enhances the CSV record pruning process by integrating a package filter to exclude certain records based on predefined patterns. This change aims to improve data accuracy and efficiency in handling CSV records.
- `CsvRecordStorageService.cs`: Introduced a package filter to prune CSV records and refactored duplicate record rejection logic.
- `PackageFilter.cs`: Added methods to filter CSV records and determine if package commit records should be ignored.
- `IPackageCommitRecord.cs`: Introduced a new interface for package commit records.
- `CatalogLeafItemRecord.cs`: Implemented the `IPackageCommitRecord` interface to support the new filtering logic.
- `PackageAssetToCsvIntegrationTest.cs`: Added a test to verify the pruning of CSV records for ignored packages.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->