Merged PR 632947: Handle bad ZIP entry in PackageContentToCsv
We handle bad entries in the other two drivers that open ZIP files (PackageFileToCsv, PackageAssemblyToCsv) but not here. This is the first time a file has a file extension included by PackageContentToCsv (`.props`) and is a bad ZIP entry.

----
Bug fix

This pull request addresses an issue with handling invalid ZIP entries in the `PackageContentToCsv` process.
- In `PackageContentToCsvDriver.cs`, added a try-catch block to handle `InvalidDataException` and log invalid ZIP entries.
- Updated `PackageContentResultType.cs` to include a new result type `InvalidZipEntry`.
- Added a test case in `PackageContentToCsvDriverTest.cs` to verify handling of invalid ZIP entries.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->